Correct package name after relocation
