+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/expressions/cast.scala
@@ -18,9 +18,9 @@
 package org.apache.flink.table.planner.expressions
 
 import org.apache.flink.api.common.typeinfo.TypeInformation
+import org.apache.flink.table.planner.typeutils.TypeCoercion
+import org.apache.flink.table.planner.validate._
 import org.apache.flink.table.runtime.types.TypeInfoLogicalTypeConverter.fromTypeInfoToLogicalType
-import org.apache.flink.table.typeutils.TypeCoercion
-import org.apache.flink.table.validate._
 
 case class Cast(child: PlannerExpression, resultType: TypeInformation[_])
   extends UnaryExpression {
