+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/expressions/composite.scala
@@ -21,7 +21,7 @@ package org.apache.flink.table.planner.expressions
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.api.common.typeutils.CompositeType
 import org.apache.flink.table.api.UnresolvedException
-import org.apache.flink.table.validate.{ValidationFailure, ValidationResult, ValidationSuccess}
+import org.apache.flink.table.planner.validate.{ValidationFailure, ValidationResult, ValidationSuccess}
 
 /**
   * Flattening of composite types. All flattenings are resolved into
