+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/validate/ValidationResult.scala
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.flink.table.validate
+package org.apache.flink.table.planner.validate
 
 /**
   * Represents the result of a validation.
